Extract formatting methods to module
